+++ src/Vish/Graphics/Data/Texture.hs
@@ -9,9 +9,9 @@ import qualified Graphics.Rendering.OpenGL.GL as GL
 data Texture
   = Texture
   { _path     :: String
-  , _srcSize  :: V2 Float
-  , _position :: V2 Float
-  , _size     :: V2 Float
+  , _srcSize  :: V2 Double
+  , _position :: V2 Double
+  , _size     :: V2 Double
   , _object   :: IORef (Maybe GL.TextureObject)
   }
